api2/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
mongoengine==0.27.0
pymongo==4.6.0
django-prometheus==2.3.1
psutil==5.9.0
